chore: refactor. introduce cpu folder with helpers
